+++ src/i18n/resources.js
@@ -1,6 +1,3 @@
 export default {
-  en: {
-    ...require('./en.json'),
-    formats: require('./en.formats.json'),
-  },
+  en: { ...require('./en.json') },
 }