+++ src/vue/pages/Profile.vue
@@ -1,0 +1,9 @@
+<template>
+  <div class="profile">
+    <!--  -->
+  </div>
+</template>
+
+<script>
+export default { name: 'profile' }
+</script>